AudioTrack: Don't tear down sink for invalid offload track
Prevent the tear down of the audio sink in client when
setplaybackrate is called on direct pcm track and fallback
to deep buffer must take place. This is necessary because
once the sink is torn down, there is no information that the
previous track was direct pcm and needs to fall back. This
would result in a continuous creation and tear down of a
direct pcm track by the client.

CRs-Fixed: 1073831
Change-Id: Ib39b5673ec4456d7b747a5ad108a1fc90f7f50de
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>